Refactor paused scene by removing unused methods and consolidating functionality
- Deleted obsolete methods related to input handling, animations, and rendering in the paused scene, streamlining the codebase.
- Removed the entire pause directory, consolidating pause-related functionality into the main paused.go file for better organization.
- Enhanced clarity and maintainability by eliminating dead code and focusing on essential features for the paused scene.
